Reservation Minor updates
**createReservation.js**
- Changed thef file name
- Changed targeted collection from trialReservation to reservation
- Added Error handling for non-inputted values

**Other**
- Fixed conflicts in .json files
- Organised App.js file a bit
- Changed link to Order in NavBar.js
