+++ client/src/Components/createReservation/createReservation.css
@@ -1,0 +1,11 @@
+.img {
+  height: max-content;
+  width: 100%;
+}
+
+.resError {
+  color: red;
+  font-size: 10px;
+  font-weight: bold;
+  margin-bottom: 10px;
+}
