+++ client/src/Components/createReservation/createReservation.js
@@ -1,0 +1,162 @@
+import React, { Component } from "react";
+import { Card, Grid, FormControl, Button } from '@material-ui/core';
+import firebase from '../../config/firebase';
+
+import backgroundImg from '../../Images/headerImage.png';
+import './createReservation.css';
+
+class CreateReservation extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      userID: '',
+      memberFullName: '',
+      numOfPpl: '',
+      dateOfRes: '',
+      timeOfRes: '',
+      additionalComms: '',
+      timestamp: '',
+      resStatus: false,
+      numPplError: '',
+      dateError: '',
+      timeError: ''
+    };
+  }
+
+  handleChange = (e) => {
+    const itemName = e.target.name;
+    const itemValue = e.target.value;
+
+    this.setState({ [itemName]: itemValue });
+  };
+
+  validate = () => {
+    let numPplError = '';
+    let dateError = '';
+    let timeError = '';
+    
+    var date = new Date();
+    var today = Date.parse(date);
+    var selectedDate = Date.parse(this.state.dateOfRes);
+
+    if (!this.state.numOfPpl) {
+      numPplError = 'Please select number of people attending';
+    }
+
+    if (!this.state.timeOfRes) {
+      timeError = 'Please indicate time of booking';
+    }
+
+    if ((selectedDate < today) || (!this.state.dateOfRes)) {
+      dateError = 'Invalid date selected'
+    }
+
+    if (numPplError || timeError || dateError) {
+      this.setState({ numPplError, timeError, dateError });
+      return false;
+    }
+
+    return true;
+  };
+
+  addReservation = (e) => {
+    e.preventDefault();
+    var isValid = this.validate();
+    const resDb = firebase.firestore();
+    var userID = firebase.auth().currentUser.uid;
+    var time = firebase.firestore.FieldValue.serverTimestamp();
+    resDb.settings({
+      timestampsInSnapshots: true
+    });
+
+    if (isValid) {
+      resDb.collection('reservations')
+        .orderBy('id', 'desc').limit(1).get().then(querySnapshot => {
+          querySnapshot.forEach(documentSnapshot => {
+            var newID = documentSnapshot.id;
+            console.log(`Found document at ${documentSnapshot.ref.path}`);
+            console.log(`Document's ID: ${documentSnapshot.id}`);
+            var newvalue = parseInt(newID, 10) + 1;
+            var ToString = "" + newvalue;
+            return resDb.collection('reservations').doc(ToString).set({
+              id: newvalue,
+              userID: userID,
+              memberFullName: this.state.memberFullName,
+              numOfPpl: this.state.numOfPpl,
+              dateOfRes: this.state.dateOfRes,
+              timeOfRes: this.state.timeOfRes,
+              additionalComms: this.state.additionalComms,
+              timestamp: time,
+              resStatus: true
+            });
+          })
+        }).then(() => {
+          window.location = 'order'
+        })
+    }
+  };
+
+  render() {
+    return (
+      <Grid container>
+        <img class="img" src={backgroundImg}></img>
+        <Grid item xs={6}>
+          <form class="reservation" onSubmit={this.addReservation} noValidate>
+            <FormControl>
+              <Card className="reservation_Info" variant="outlined">
+                <h1>Reservation</h1>
+                <label for="numOfPpl" class="numOfPplLbl"><b>No. of People</b></label>
+                <br />
+                <input type="number" min="1" max="12" name="numOfPpl" class="numOfPpl" required value={this.state.numOfPpl} onChange={this.handleChange} />
+                <div class="resError">
+                  {this.state.numPplError}
+                </div>
+                <label for="dateofRes" class="dateLbl"><b>Date</b></label>
+                <br />
+                <input type="date" name="dateOfRes" class="date" required value={this.state.dateOfRes} onChange={this.handleChange} />
+                <div class="resError">
+                {this.state.dateError}
+                </div>
+                <label for="timeOfRes" class="timeLbl"><b>Time</b></label>
+                <br />
+                <select name="timeOfRes" class="time" required value={this.state.timeOfRes} onChange={this.handleChange}>
+                  <option selected value="" disabled hidden />
+                  <option value="lunchTimes" disabled>Lunch</option>
+                  <option value="11am">11:00</option>
+                  <option value="11.30am">11:30</option>
+                  <option value="12pm">12:00</option>
+                  <option value="12.30pm">12:30</option>
+                  <option value="1pm">13:00</option>
+                  <option value="1.30pm">13:30</option>
+                  <option value="2pm">14:00</option>
+                  <option value="2.30pm">14:30</option>
+                  <option value="dinnerTimes" disabled>Dinner</option>
+                  <option value="5pm">17:00</option>
+                  <option value="5.30pm">17:30</option>
+                  <option value="6pm">18:00</option>
+                  <option value="6.30pm">18:30</option>
+                  <option value="7pm">19:00</option>
+                  <option value="7.30pm">19:30</option>
+                  <option value="8pm">20:00</option>
+                  <option value="8.30pm">20:30</option>
+                  <option value="9pm">21:00</option>
+                </select>
+                <div class="resError">
+                {this.state.timeError}
+                </div>
+                <label for="additionalComms" class="commentsLbl"><b>Additional Comments</b></label>
+                <br />
+                <input type="text" name="additionalComms" class="comments" value={this.state.additionalComms} onChange={this.handleChange} />
+                <br></br>
+                <br></br>
+                <button>Book</button>
+              </Card>
+            </FormControl>
+          </form>
+        </Grid>
+      </Grid >
+    );
+  }
+}
+
+export default CreateReservation;
