+++ package.json
@@ -23,6 +23,7 @@
     "generate-password": "^1.5.1",
     "react-bootstrap": "^1.0.1",
     "react-bootstrap-validation": "^0.1.11",
+    "react-confirm-alert": "^2.6.1",
     "react-image": "^2.4.0",
     "react-router-dom": "^5.2.0",
     "react-star-rating-component": "^1.4.1"
